removed buildIdMethods function and renaming
