+++ src/collection/objects/aiModel.js
@@ -1,4 +1,4 @@
-const Component = require('./_component');
+const Component = require('./_objects');
 const {
   get,
 } = require('./_util');
